admin/pyproject.toml
[project]
name = "admin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.0",
    "fastapi[standard]>=0.115.6",
    "ffmpeg-python>=0.2.0",
    "jinja2>=3.1.5",
    "numpy==2.0.2",
    "passlib>=1.7.4",
    "pyhumps>=3.8.0",
    "pyjwt>=2.10.1",
    "sqlalchemy-utils>=0.41.2",
    "sqlmodel>=0.0.22",
    "torch>=2.5.1",
    "transformers>=4.48.0",
]

[dependency-groups]
dev = [
    "ruff>=0.9.0",
]
